chore: convert commons.dom to ES Modules
