+++ .eslintrc
@@ -76,6 +76,7 @@
       "files": [
         "lib/commons/aria/*.js",
         "lib/commons/color/*.js",
+        "lib/commons/dom/*.js",
         "lib/commons/forms/*.js",
         "lib/commons/matches/*.js",
         "lib/commons/table/*.js",
