+++ lib/commons/aria/arialabelledby-text.js
@@ -1,4 +1,5 @@
-/* global dom, text */
+import idrefs from '../dom/idrefs';
+import accessibleText from '../text/accessible-text';
 
 /**
  * Get the accessible name based on aria-labelledby
@@ -34,11 +35,9 @@ function arialabelledbyText(node, context = {}) {
 		return '';
 	}
 
-	// TODO: es-module-dom.idrefs
-	const refs = dom.idrefs(node, 'aria-labelledby').filter(elm => elm);
+	const refs = idrefs(node, 'aria-labelledby').filter(elm => elm);
 	return refs.reduce((accessibleName, elm) => {
-		// TODO: es-module-text.accessibleText
-		const accessibleNameAdd = text.accessibleText(elm, {
+		const accessibleNameAdd = accessibleText(elm, {
 			// Prevent the infinite reference loop:
 			inLabelledByContext: true,
 			startNode: context.startNode || node,
