+++ lib/commons/aria/get-owned-virtual.js
@@ -1,4 +1,4 @@
-/* global dom */
+import idrefs from '../dom/idrefs';
 
 /**
  * Get an element's owned elements
@@ -14,8 +14,7 @@ function getOwnedVirtual({ actualNode, children }) {
 	// TODO: Descend into children with role=presentation|none
 	// TODO: Exclude descendents owned by other elements
 
-	// TODO: es-module-dom.idrefs
-	return dom.idrefs(actualNode, 'aria-owns').reduce((ownedElms, element) => {
+	return idrefs(actualNode, 'aria-owns').reduce((ownedElms, element) => {
 		if (element) {
 			// TODO: es-module-utils.getNodeFromTree
 			const virtualNode = axe.utils.getNodeFromTree(element);
