+++ lib/commons/aria/is-accessible-ref.js
@@ -1,5 +1,5 @@
-/* global dom */
 import lookupTable from './lookup-table';
+import getRootNode from '../dom/get-root-node';
 
 const idRefsRegex = /^idrefs?$/;
 
@@ -42,8 +42,7 @@ function cacheIdRefs(node, refAttrs) {
  */
 function isAccessibleRef(node) {
 	node = node.actualNode || node;
-	// TODO: es-module-dom.getRootNode
-	let root = dom.getRootNode(node);
+	let root = getRootNode(node);
 	root = root.documentElement || root; // account for shadow roots
 	const id = node.id;
