+++ lib/commons/aria/label-virtual.js
@@ -1,4 +1,6 @@
-/* global dom, text */
+import idrefs from '../dom/idrefs';
+import visibleVirtual from '../text/visible-virtual';
+import sanitize from '../text/sanitize';
 
 /**
  * Gets the accessible ARIA label text of a given element
@@ -14,14 +16,12 @@ function labelVirtual({ actualNode }) {
 
 	if (actualNode.getAttribute('aria-labelledby')) {
 		// aria-labelledby
-		// TODO: es-module-dom.idrefs
-		ref = dom.idrefs(actualNode, 'aria-labelledby');
+		ref = idrefs(actualNode, 'aria-labelledby');
 		candidate = ref
 			.map(function(thing) {
 				// TODO: es-module-utils.getNodeFromTree
 				const vNode = axe.utils.getNodeFromTree(thing);
-				// TODO: es-module-text.visibleVirtual
-				return vNode ? text.visibleVirtual(vNode, true) : '';
+				return vNode ? visibleVirtual(vNode, true) : '';
 			})
 			.join(' ')
 			.trim();
@@ -34,8 +34,7 @@ function labelVirtual({ actualNode }) {
 	// aria-label
 	candidate = actualNode.getAttribute('aria-label');
 	if (candidate) {
-		// TODO: es-module-text.sanitize
-		candidate = text.sanitize(candidate).trim();
+		candidate = sanitize(candidate).trim();
 		if (candidate) {
 			return candidate;
 		}
