+++ lib/commons/aria/validate-attr-value.js
@@ -1,5 +1,5 @@
-/* global dom */
 import lookupTable from './lookup-table';
+import getRootNode from '../dom/get-root-node';
 
 /**
  * Validate the value of an ARIA attribute
@@ -16,9 +16,7 @@ function validateAttrValue(node, attr) {
 	let list;
 	const value = node.getAttribute(attr);
 	const attrInfo = lookupTable.attributes[attr];
-
-	// TODO: es-module-dom.getRootNode
-	var doc = dom.getRootNode(node);
+	const doc = getRootNode(node);
 
 	if (!attrInfo) {
 		return true;
