+++ lib/commons/color/get-background-color.js
@@ -1,10 +1,10 @@
-/* global dom */
 import incompleteData from './incomplete-data';
 import getBackgroundStack from './get-background-stack';
 import getOwnBackgroundColor from './get-own-background-color';
 import elementHasImage from './element-has-image';
 import Color from './color';
 import flattenColors from './flatten-colors';
+import visuallyContains from '../dom/visually-contains';
 
 /**
  * Determine if element is partially overlapped, triggering a Can't Tell result
@@ -16,8 +16,7 @@ import flattenColors from './flatten-colors';
  */
 function elmPartiallyObscured(elm, bgElm, bgColor) {
 	var obscured =
-		// TODO: es-module-dom.visuallyContains
-		elm !== bgElm && !dom.visuallyContains(elm, bgElm) && bgColor.alpha !== 0;
+		elm !== bgElm && !visuallyContains(elm, bgElm) && bgColor.alpha !== 0;
 	if (obscured) {
 		incompleteData.set('bgColor', 'elmPartiallyObscured');
 	}
