+++ lib/commons/color/get-background-stack.js
@@ -1,9 +1,9 @@
-/* global dom */
-
 import filteredRectStack from './filtered-rect-stack';
 import elementHasImage from './element-has-image';
 import getOwnBackgroundColor from './get-own-background-color';
 import incompleteData from './incomplete-data';
+import shadowElementsFromPoint from '../dom/shadow-elements-from-point';
+import reduceToElementsBelowFloating from '../dom/reduce-to-elements-below-floating';
 
 /**
  * Determines overlap of node's content with a bgNode. Used for inline elements
@@ -16,8 +16,7 @@ function contentOverlapping(targetElement, bgNode) {
 	// get content box of target element
 	// check to see if the current bgNode is overlapping
 	var targetRect = targetElement.getClientRects()[0];
-	// TODO: es-module-dom.shadowElementsFromPoint
-	var obscuringElements = dom.shadowElementsFromPoint(
+	var obscuringElements = shadowElementsFromPoint(
 		targetRect.left,
 		targetRect.top
 	);
@@ -115,8 +114,7 @@ function getBackgroundStack(elm) {
 	if (elmStack === null) {
 		return null;
 	}
-	// TODO: es-module-dom.reduceToElementsBelowFloating
-	elmStack = dom.reduceToElementsBelowFloating(elmStack, elm);
+	elmStack = reduceToElementsBelowFloating(elmStack, elm);
 	elmStack = sortPageBackground(elmStack);
 
 	// Return all elements BELOW the current element, null if the element is undefined
