+++ lib/commons/color/get-rect-stack.js
@@ -1,3 +1,6 @@
+import getElementStack from '../dom/get-element-stack';
+import getTextElementStack from '../dom/get-text-element-stack';
+
 /**
  * Get relevant stacks of block and inline elements, excluding line breaks
  * @method getRectStack
@@ -6,12 +9,10 @@
  * @return {Array}
  */
 function getRectStack(elm) {
-	// TODO: es-module-dom.getElementStack
-	const boundingStack = axe.commons.dom.getElementStack(elm);
+	const boundingStack = getElementStack(elm);
 
 	// Handle inline elements spanning multiple lines to be evaluated
-	// TODO: es-module-dom.getTextElementStack
-	const filteredArr = axe.commons.dom.getTextElementStack(elm);
+	const filteredArr = getTextElementStack(elm);
 
 	// If the element does not have multiple rects, like for display:block, return a single stack
 	if (!filteredArr || filteredArr.length <= 1) {
