+++ lib/commons/dom/find-elms-in-context.js
@@ -1,4 +1,4 @@
-/* global axe, dom */
+import getRootNode from './get-root-node';
 
 /**
  * Find elements referenced from a given context
@@ -12,17 +12,20 @@
  * @param {String} element.elm NodeName to search for (optional)
  * @return {Array<Node>}
  */
-dom.findElmsInContext = function({ context, value, attr, elm = '' }) {
+function findElmsInContext({ context, value, attr, elm = '' }) {
 	let root;
+	// TODO: es-module-utils.escapeSelector
 	const escapedValue = axe.utils.escapeSelector(value);
 
 	if (context.nodeType === 9 || context.nodeType === 11) {
 		// It's already root
 		root = context;
 	} else {
-		root = dom.getRootNode(context);
+		root = getRootNode(context);
 	}
 	return Array.from(
 		root.querySelectorAll(elm + '[' + attr + '=' + escapedValue + ']')
 	);
-};
+}
+
+export default findElmsInContext;
