+++ lib/commons/dom/get-composed-parent.js
@@ -1,4 +1,3 @@
-/* global dom */
 /**
  * Get an element's parent in the flattened tree
  * @method getComposedParent
@@ -7,7 +6,7 @@
  * @param {Node} element
  * @return {Node|null} Parent element or Null for root node
  */
-dom.getComposedParent = function getComposedParent(element) {
+function getComposedParent(element) {
 	if (element.assignedSlot) {
 		// NOTE: If the display of a slot element isn't 'contents',
 		// the slot shouldn't be ignored. Chrome does not support this (yet) so,
@@ -22,4 +21,6 @@ dom.getComposedParent = function getComposedParent(element) {
 		}
 	}
 	return null; // Root node
-};
+}
+
+export default getComposedParent;
