+++ lib/commons/dom/get-element-by-reference.js
@@ -1,5 +1,3 @@
-/* global dom */
-
 /**
  * Returns a reference to the element matching the attr URL fragment value
  * @method getElementByReference
@@ -9,7 +7,7 @@
  * @param {String} attr Attribute name (href)
  * @return {Element}
  */
-dom.getElementByReference = function(node, attr) {
+function getElementByReference(node, attr) {
 	let fragment = node.getAttribute(attr);
 	if (!fragment) {
 		return null;
@@ -31,4 +29,6 @@ dom.getElementByReference = function(node, attr) {
 		return candidate[0];
 	}
 	return null;
-};
+}
+
+export default getElementByReference;
