+++ lib/commons/dom/get-element-coordinates.js
@@ -1,4 +1,4 @@
-/* global dom */
+import getScrollOffset from './get-scroll-offset';
 
 /**
  * Get the coordinates of the element passed into the function relative to the document
@@ -19,10 +19,10 @@
  * @property {Number} width The width of the element
  * @property {Number} height The height of the element
  */
-dom.getElementCoordinates = function(element) {
+function getElementCoordinates(element) {
 	'use strict';
 
-	var scrollOffset = dom.getScrollOffset(document),
+	var scrollOffset = getScrollOffset(document),
 		xOffset = scrollOffset.left,
 		yOffset = scrollOffset.top,
 		coords = element.getBoundingClientRect();
@@ -35,4 +35,6 @@ dom.getElementCoordinates = function(element) {
 		width: coords.right - coords.left,
 		height: coords.bottom - coords.top
 	};
-};
+}
+
+export default getElementCoordinates;
