+++ lib/commons/dom/get-root-node.js
@@ -1,5 +1,3 @@
-/* global dom */
-
 /**
  * Return the document or document fragment (shadow DOM)
  * @method getRootNode
@@ -9,4 +7,5 @@
  * @returns {DocumentFragment|Document}
  * @deprecated use axe.utils.getRootNode
  */
-dom.getRootNode = axe.utils.getRootNode;
+// TODO: es-module-utils.getRootNode
+export default axe.utils.getRootNode;
