+++ lib/commons/dom/get-scroll-offset.js
@@ -1,4 +1,3 @@
-/* global dom */
 /**
  * Get the scroll offset of the document passed in
  * @method getScrollOffset
@@ -7,7 +6,7 @@
  * @param {Document} element The element to evaluate, defaults to document
  * @return {Object} Contains the attributes `x` and `y` which contain the scroll offsets
  */
-dom.getScrollOffset = function(element) {
+function getScrollOffset(element) {
 	'use strict';
 
 	if (!element.nodeType && element.document) {
@@ -30,4 +29,6 @@ dom.getScrollOffset = function(element) {
 		left: element.scrollLeft,
 		top: element.scrollTop
 	};
-};
+}
+
+export default getScrollOffset;
