+++ lib/commons/dom/get-tabbable-elements.js
@@ -1,5 +1,3 @@
-/* global dom */
-
 /**
  * Get all elements (including given node) that are part of the tab order
  * @method getTabbableElements
@@ -8,7 +6,8 @@
  * @param  {Object} virtualNode The virtualNode to assess
  * @return {Boolean}
  */
-dom.getTabbableElements = function getTabbableElements(virtualNode) {
+function getTabbableElements(virtualNode) {
+	// TODO: es-module-utils.querySelectorAll
 	const nodeAndDescendents = axe.utils.querySelectorAll(virtualNode, '*');
 
 	const tabbableElements = nodeAndDescendents.filter(vNode => {
@@ -21,4 +20,6 @@ dom.getTabbableElements = function getTabbableElements(virtualNode) {
 	});
 
 	return tabbableElements;
-};
+}
+
+export default getTabbableElements;
