+++ lib/commons/dom/get-viewport-size.js
@@ -1,4 +1,3 @@
-/* global dom */
 /**
  * Gets the width and height of the viewport; used to calculate the right and bottom boundaries of the viewable area.
  * @method getViewportSize
@@ -7,12 +6,11 @@
  * @param  {Object}  win The `window` object that should be measured
  * @return {Object}  Object with the `width` and `height` of the viewport
  */
-dom.getViewportSize = function(win) {
+function getViewportSize(win) {
 	'use strict';
 
-	var body,
-		doc = win.document,
-		docElement = doc.documentElement;
+	const doc = win.document;
+	const docElement = doc.documentElement;
 
 	if (win.innerWidth) {
 		return {
@@ -28,10 +26,12 @@ dom.getViewportSize = function(win) {
 		};
 	}
 
-	body = doc.body;
+	const body = doc.body;
 
 	return {
 		width: body.clientWidth,
 		height: body.clientHeight
 	};
-};
+}
+
+export default getViewportSize;
