+++ lib/commons/dom/has-content-virtual.js
@@ -1,4 +1,6 @@
-/* global dom, aria, axe */
+import isVisualContent from './is-visual-content';
+import labelVirtual from '../aria/label-virtual';
+
 const hiddenTextElms = [
 	'HEAD',
 	'TITLE',
@@ -32,44 +34,20 @@ function hasChildTextNodes(elm) {
  * @param {Boolean} ignoreAria if true, ignores `aria label` computation for content deduction
  * @return {Boolean}
  */
-dom.hasContentVirtual = function(elm, noRecursion, ignoreAria) {
+function hasContentVirtual(elm, noRecursion, ignoreAria) {
 	return (
 		// It has text
 		hasChildTextNodes(elm) ||
 		// It is a graphical element
-		dom.isVisualContent(elm.actualNode) ||
+		isVisualContent(elm.actualNode) ||
 		// It has an ARIA label
-		(!ignoreAria && !!aria.labelVirtual(elm)) ||
+		(!ignoreAria && !!labelVirtual(elm)) ||
 		// or one of it's descendants does
 		(!noRecursion &&
 			elm.children.some(
-				child => child.actualNode.nodeType === 1 && dom.hasContentVirtual(child)
+				child => child.actualNode.nodeType === 1 && hasContentVirtual(child)
 			))
 	);
-};
+}
 
-/**
- * Find virtual node and call hasContentVirtual()
- * IMPORTANT: This method requires the composed tree at axe._tree
- * @see axe.commons.dom.hasContentVirtual
- * @method hasContent
- * @memberof axe.commons.dom
- * @instance
- * @param {DOMNode} elm DOMNode element to check
- * @param	{Boolean} noRecursion If true, only the element is checked, otherwise it will search all child nodes
- * @param {Boolean} ignoreAria if true, ignores `aria label` computation for content deduction
- * @return {Boolean}
- */
-dom.hasContent = function hasContent(elm, noRecursion, ignoreAria) {
-	elm = axe.utils.getNodeFromTree(elm);
-	return dom.hasContentVirtual(elm, noRecursion, ignoreAria);
-};
-
-/**
- * @typedef VirtualNode
- * @type {Object}
- * @property {HTMLElement} actualNode reference to actual HTML DOM element
- * @property {Array<VirtualNode>} children array of child virtualNodes
- * @property {String} shadowId The shadowId indicates whether the node is in a shadow root
- * and if it is, which one it is inside the boundary.
- */
+export default hasContentVirtual;
