+++ lib/commons/dom/idrefs.js
@@ -1,4 +1,4 @@
-/*global dom, axe */
+import getRootNode from './get-root-node';
 
 /**
  * Get elements referenced via a space-separated token attribute;
@@ -16,21 +16,22 @@
  * on the flattened tree to dereference idrefs.
  *
  */
-dom.idrefs = function(node, attr) {
+function idrefs(node, attr) {
 	'use strict';
 
-	var index,
-		length,
-		doc = dom.getRootNode(node),
-		result = [],
-		idrefs = node.getAttribute(attr);
+	const doc = getRootNode(node);
+	const result = [];
+	let attrValue = node.getAttribute(attr);
 
-	if (idrefs) {
-		idrefs = axe.utils.tokenList(idrefs);
-		for (index = 0, length = idrefs.length; index < length; index++) {
-			result.push(doc.getElementById(idrefs[index]));
+	if (attrValue) {
+		// TODO: es-module-utils.tokenList
+		attrValue = axe.utils.tokenList(attrValue);
+		for (let index = 0; index < attrValue.length; index++) {
+			result.push(doc.getElementById(attrValue[index]));
 		}
 	}
 
 	return result;
-};
+}
+
+export default idrefs;
