+++ lib/commons/dom/is-hidden-with-css.js
@@ -1,4 +1,4 @@
-/* global dom */
+import getComposedParent from './get-composed-parent';
 
 /**
  * Determine whether an element is hidden based on css
@@ -9,7 +9,8 @@
  * @param {Boolean} descendentVisibilityValue (Optional) immediate descendant visibility value used for recursive computation
  * @return {Boolean} the element's hidden status
  */
-dom.isHiddenWithCSS = function isHiddenWithCSS(el, descendentVisibilityValue) {
+function isHiddenWithCSS(el, descendentVisibilityValue) {
+	// TODO: es-module-utils.getNodeFromTree
 	const vNode = axe.utils.getNodeFromTree(el);
 
 	if (!vNode) {
@@ -21,7 +22,7 @@ dom.isHiddenWithCSS = function isHiddenWithCSS(el, descendentVisibilityValue) {
 	}
 
 	return vNode._isHiddenWithCSS;
-};
+}
 
 function _isHiddenWithCSS(el, descendentVisibilityValue) {
 	if (el.nodeType === 9) {
@@ -65,9 +66,11 @@ function _isHiddenWithCSS(el, descendentVisibilityValue) {
 		return true;
 	}
 
-	const parent = dom.getComposedParent(el);
+	const parent = getComposedParent(el);
 	if (parent && !HIDDEN_VISIBILITY_VALUES.includes(visibilityValue)) {
-		return dom.isHiddenWithCSS(parent, visibilityValue);
+		return isHiddenWithCSS(parent, visibilityValue);
 	}
 	return false;
 }
+
+export default isHiddenWithCSS;
