+++ lib/commons/dom/is-html5.js
@@ -1,5 +1,3 @@
-/* global dom */
-
 /**
  * Determines if a document node is HTML 5
  * @method isHTML5
@@ -8,10 +6,12 @@
  * @param {Node} doc
  * @return {Boolean}
  */
-dom.isHTML5 = function(doc) {
-	var node = doc.doctype;
+function isHTML5(doc) {
+	const node = doc.doctype;
 	if (node === null) {
 		return false;
 	}
 	return node.name === 'html' && !node.publicId && !node.systemId;
-};
+}
+
+export default isHTML5;
