+++ lib/commons/dom/is-in-text-block.js
@@ -1,11 +1,20 @@
-/* global axe, dom, window */
+import getComposedParent from './get-composed-parent';
+import sanitize from '../text/sanitize';
+
 function walkDomNode(node, functor) {
 	if (functor(node.actualNode) !== false) {
 		node.children.forEach(child => walkDomNode(child, functor));
 	}
 }
 
-var blockLike = ['block', 'list-item', 'table', 'flex', 'grid', 'inline-block'];
+const blockLike = [
+	'block',
+	'list-item',
+	'table',
+	'flex',
+	'grid',
+	'inline-block'
+];
 function isBlock(elm) {
 	var display = window.getComputedStyle(elm).getPropertyValue('display');
 	return blockLike.includes(display) || display.substr(0, 6) === 'table-';
@@ -13,10 +22,12 @@ function isBlock(elm) {
 
 function getBlockParent(node) {
 	// Find the closest parent
-	let parentBlock = dom.getComposedParent(node);
+	let parentBlock = getComposedParent(node);
 	while (parentBlock && !isBlock(parentBlock)) {
-		parentBlock = dom.getComposedParent(parentBlock);
+		parentBlock = getComposedParent(parentBlock);
 	}
+
+	// TODO: es-module-utils.getNodeFromTree
 	return axe.utils.getNodeFromTree(parentBlock);
 }
 
@@ -25,7 +36,7 @@ function getBlockParent(node) {
  * @param  {Element} node [description]
  * @return {Boolean}      [description]
  */
-dom.isInTextBlock = function isInTextBlock(node) {
+function isInTextBlock(node) {
 	if (isBlock(node)) {
 		// Ignore if the link is a block
 		return false;
@@ -87,8 +98,10 @@ dom.isInTextBlock = function isInTextBlock(node) {
 		}
 	});
 
-	parentText = axe.commons.text.sanitize(parentText);
-	linkText = axe.commons.text.sanitize(linkText);
+	parentText = sanitize(parentText);
+	linkText = sanitize(linkText);
 
 	return parentText.length > linkText.length;
-};
+}
+
+export default isInTextBlock;
