+++ lib/commons/dom/is-modal-open.js
@@ -1,4 +1,5 @@
-/* global dom, axe */
+import isVisible from './is-visible';
+import getViewportSize from './get-viewport-size';
 
 /**
  * Determines if there is a modal currently open.
@@ -7,7 +8,7 @@
  * @instance
  * @return {Boolean|undefined} True if we know (or our best guess) that a modal is open, undefined if we can't tell (doesn't mean there isn't one open)
  */
-dom.isModalOpen = function isModalOpen(options) {
+function isModalOpen(options) {
 	options = options || {};
 	let modalPercent = options.modalPercent || 0.75;
 
@@ -31,17 +32,21 @@ dom.isModalOpen = function isModalOpen(options) {
 	// in the viewport. since we aren't sure if it is or is not a modal this is
 	// just our best guess of being one based on convention.
 
+	// TODO: es-module-_cache
 	if (axe._cache.get('isModalOpen')) {
 		return axe._cache.get('isModalOpen');
 	}
 
+	// TODO: es-module-utils.querySelectorAllFilter
 	const definiteModals = axe.utils.querySelectorAllFilter(
+		// TODO: es-module-_tree
 		axe._tree[0],
 		'dialog, [role=dialog], [aria-modal=true]',
-		vNode => dom.isVisible(vNode.actualNode)
+		vNode => isVisible(vNode.actualNode)
 	);
 
 	if (definiteModals.length) {
+		// TODO: es-module-_cache
 		axe._cache.set('isModalOpen', true);
 		return true;
 	}
@@ -50,7 +55,7 @@ dom.isModalOpen = function isModalOpen(options) {
 	// four corners and one from the middle of the viewport (total of 5). if each
 	// stack contains an element whose width/height is >= 75% of the screen, we
 	// found a "could be a modal"
-	const viewport = dom.getViewportSize(window);
+	const viewport = getViewportSize(window);
 	const percentWidth = viewport.width * modalPercent;
 	const percentHeight = viewport.height * modalPercent;
 	const x = (viewport.width - percentWidth) / 2;
@@ -88,11 +93,15 @@ dom.isModalOpen = function isModalOpen(options) {
 		});
 
 		if (modalElement && stacks.every(stack => stack.includes(modalElement))) {
+			// TODO: es-module-_cache
 			axe._cache.set('isModalOpen', true);
 			return true;
 		}
 	}
 
+	// TODO: es-module-_cache
 	axe._cache.set('isModalOpen', undefined);
 	return undefined;
-};
+}
+
+export default isModalOpen;
