+++ lib/commons/dom/is-node.js
@@ -1,13 +1,15 @@
-/* global dom */
 /**
  * Determines if element is an instance of Node
  * @method isNode
  * @memberof axe.commons.dom
  * @instance
+ * @deprecated
  * @param  {Element} element
  * @return {Boolean}
  */
-dom.isNode = function(element) {
+function isNode(element) {
 	'use strict';
 	return element instanceof Node;
-};
+}
+
+export default isNode;
