+++ lib/commons/dom/is-offscreen.js
@@ -1,7 +1,9 @@
-/* global dom */
+import getComposedParent from './get-composed-parent';
+import getElementCoordinates from './get-element-coordinates';
+import getViewportSize from './get-viewport-size';
 
 function noParentScrolled(element, offset) {
-	element = dom.getComposedParent(element);
+	element = getComposedParent(element);
 	while (element && element.nodeName.toLowerCase() !== 'html') {
 		if (element.scrollTop) {
 			offset += element.scrollTop;
@@ -9,7 +11,7 @@ function noParentScrolled(element, offset) {
 				return false;
 			}
 		}
-		element = dom.getComposedParent(element);
+		element = getComposedParent(element);
 	}
 	return true;
 }
@@ -22,14 +24,14 @@ function noParentScrolled(element, offset) {
  * @param  {Element} element
  * @return {Boolean}
  */
-dom.isOffscreen = function(element) {
+function isOffscreen(element) {
 	let leftBoundary;
 	const docElement = document.documentElement;
 	const styl = window.getComputedStyle(element);
 	const dir = window
 		.getComputedStyle(document.body || docElement)
 		.getPropertyValue('direction');
-	const coords = dom.getElementCoordinates(element);
+	const coords = getElementCoordinates(element);
 
 	// bottom edge beyond
 	if (
@@ -51,7 +53,7 @@ dom.isOffscreen = function(element) {
 	} else {
 		leftBoundary = Math.max(
 			docElement.scrollWidth,
-			dom.getViewportSize(window).width
+			getViewportSize(window).width
 		);
 		if (coords.left >= leftBoundary) {
 			return true;
@@ -59,4 +61,6 @@ dom.isOffscreen = function(element) {
 	}
 
 	return false;
-};
+}
+
+export default isOffscreen;
