+++ lib/commons/dom/is-skip-link.js
@@ -1,5 +1,3 @@
-/* global dom */
-
 // test for hrefs that start with # or /# (for angular)
 const isInternalLinkRegex = /^\/?#[^/!]/;
 
@@ -11,24 +9,28 @@ const isInternalLinkRegex = /^\/?#[^/!]/;
  * @param  {Element} element
  * @return {Boolean}
  */
-dom.isSkipLink = function(element) {
+function isSkipLink(element) {
 	if (!isInternalLinkRegex.test(element.getAttribute('href'))) {
 		return false;
 	}
 
 	let firstPageLink;
+	// TODO: es-module-_cache
 	if (typeof axe._cache.get('firstPageLink') !== 'undefined') {
 		firstPageLink = axe._cache.get('firstPageLink');
 	} else {
 		// define a skip link as any anchor element whose href starts with `#...`
 		// and which precedes the first anchor element whose href doesn't start
 		// with  `#...` (that is, a link to a page)
+		// TODO: es-module-utils.querySelectorAll
 		firstPageLink = axe.utils.querySelectorAll(
+			// TODO: es-module-_tree
 			axe._tree,
 			'a:not([href^="#"]):not([href^="/#"]):not([href^="javascript"])'
 		)[0];
 
 		// null will signify no first page link
+		// TODO: es-module-_cache
 		axe._cache.set('firstPageLink', firstPageLink || null);
 	}
 
@@ -42,4 +44,6 @@ dom.isSkipLink = function(element) {
 		element.compareDocumentPosition(firstPageLink.actualNode) ===
 		element.DOCUMENT_POSITION_FOLLOWING
 	);
-};
+}
+
+export default isSkipLink;
