+++ lib/commons/dom/is-visible.js
@@ -1,4 +1,7 @@
-/* global dom */
+import getRootNode from './get-root-node';
+import isOffscreen from './is-offscreen';
+import findUp from './find-up';
+
 const clipRegex = /rect\s*\(([0-9]+)px,?\s*([0-9]+)px,?\s*([0-9]+)px,?\s*([0-9]+)px\s*\)/;
 const clipPathRegex = /(\w+)\((\d+)/;
 
@@ -52,7 +55,7 @@ function isAreaVisible(el, screenReader, recursed) {
 	 * - Verified that `map` element cannot refer to `area` elements across different document trees
 	 * - Verified that `map` element does not get affected by altering `display` property
 	 */
-	const mapEl = dom.findUp(el, 'map');
+	const mapEl = findUp(el, 'map');
 	if (!mapEl) {
 		return false;
 	}
@@ -65,13 +68,16 @@ function isAreaVisible(el, screenReader, recursed) {
 	/**
 	 * `map` element has to be in light DOM
 	 */
-	const mapElRootNode = dom.getRootNode(el);
+	const mapElRootNode = getRootNode(el);
 	if (!mapElRootNode || mapElRootNode.nodeType !== 9) {
 		return false;
 	}
 
+	// TODO: es-module-utils.querySelectorAll
 	const refs = axe.utils.querySelectorAll(
+		// TODO: es-module-_tree
 		axe._tree,
+		// TODO: es-module-utils.escapeSelector
 		`img[usemap="#${axe.utils.escapeSelector(mapElName)}"]`
 	);
 	if (!refs || !refs.length) {
@@ -79,7 +85,7 @@ function isAreaVisible(el, screenReader, recursed) {
 	}
 
 	return refs.some(({ actualNode }) =>
-		dom.isVisible(actualNode, screenReader, recursed)
+		isVisible(actualNode, screenReader, recursed)
 	);
 }
 
@@ -93,8 +99,9 @@ function isAreaVisible(el, screenReader, recursed) {
  * @param {Boolean} recursed
  * @return {Boolean} The element's visibilty status
  */
-dom.isVisible = function(el, screenReader, recursed) {
+function isVisible(el, screenReader, recursed) {
 	'use strict';
+	// TODO: es-module-utils.getNodeFromTree
 	const node = axe.utils.getNodeFromTree(el);
 	const cacheName = '_isVisible' + (screenReader ? 'ScreenReader' : '');
 
@@ -131,7 +138,7 @@ dom.isVisible = function(el, screenReader, recursed) {
 			// visibility is only accurate on the first element
 			(style.getPropertyValue('visibility') === 'hidden' ||
 				// position does not matter if it was already calculated
-				(!screenReader && dom.isOffscreen(el)))) ||
+				(!screenReader && isOffscreen(el)))) ||
 		(screenReader && el.getAttribute('aria-hidden') === 'true')
 	) {
 		return false;
@@ -145,14 +152,16 @@ dom.isVisible = function(el, screenReader, recursed) {
 	}
 
 	const parent = el.assignedSlot ? el.assignedSlot : el.parentNode;
-	let isVisible = false;
+	let visible = false;
 	if (parent) {
-		isVisible = dom.isVisible(parent, screenReader, true);
+		visible = isVisible(parent, screenReader, true);
 	}
 
 	if (node) {
-		node[cacheName] = isVisible;
+		node[cacheName] = visible;
 	}
 
-	return isVisible;
-};
+	return visible;
+}
+
+export default isVisible;
