+++ lib/commons/dom/is-visual-content.js
@@ -1,5 +1,3 @@
-/*global dom */
-
 const visualRoles = [
 	'checkbox',
 	'img',
@@ -18,7 +16,7 @@ const visualRoles = [
  * @param  {Element} element The element to check
  * @return {Boolean}
  */
-dom.isVisualContent = function(element) {
+function isVisualContent(element) {
 	/*eslint indent: 0*/
 	const role = element.getAttribute('role');
 	if (role) {
@@ -46,4 +44,6 @@ dom.isVisualContent = function(element) {
 		default:
 			return false;
 	}
-};
+}
+
+export default isVisualContent;
