+++ lib/commons/dom/reduce-to-elements-below-floating.js
@@ -1,5 +1,3 @@
-/* global dom */
-
 /**
  * Reduce an array of elements to only those that are below a 'floating' element.
  * @method reduceToElementsBelowFloating
@@ -9,22 +7,19 @@
  * @param {Element} targetNode
  * @returns {Array}
  */
-dom.reduceToElementsBelowFloating = function(elements, targetNode) {
-	var floatingPositions = ['fixed', 'sticky'],
-		finalElements = [],
-		targetFound = false,
-		index,
-		currentNode,
-		style;
+function reduceToElementsBelowFloating(elements, targetNode) {
+	const floatingPositions = ['fixed', 'sticky'];
+	const finalElements = [];
+	let targetFound = false;
 
 	// Filter out elements that are temporarily floating above the target
-	for (index = 0; index < elements.length; ++index) {
-		currentNode = elements[index];
+	for (let index = 0; index < elements.length; ++index) {
+		const currentNode = elements[index];
 		if (currentNode === targetNode) {
 			targetFound = true;
 		}
 
-		style = window.getComputedStyle(currentNode);
+		const style = window.getComputedStyle(currentNode);
 
 		if (!targetFound && floatingPositions.indexOf(style.position) !== -1) {
 			//Target was not found yet, so it must be under this floating thing (and will not always be under it)
@@ -36,4 +31,6 @@ dom.reduceToElementsBelowFloating = function(elements, targetNode) {
 	}
 
 	return finalElements;
-};
+}
+
+export default reduceToElementsBelowFloating;
