+++ lib/commons/dom/shadow-elements-from-point.js
@@ -1,4 +1,6 @@
-/* global axe, dom */
+import getRootNode from './get-root-node';
+import visuallyContains from './visually-contains';
+
 /**
  * Get elements from point across shadow dom boundaries
  * @method shadowElementsFromPoint
@@ -9,7 +11,7 @@
  * @param {Object} [root] Shadow root or document root
  * @return {Array}
  */
-dom.shadowElementsFromPoint = function(nodeX, nodeY, root = document, i = 0) {
+function shadowElementsFromPoint(nodeX, nodeY, root = document, i = 0) {
 	if (i > 999) {
 		throw new Error('Infinite loop detected');
 	}
@@ -18,10 +20,11 @@ dom.shadowElementsFromPoint = function(nodeX, nodeY, root = document, i = 0) {
 		Array.from(root.elementsFromPoint(nodeX, nodeY) || [])
 			// As of Chrome 66, elementFromPoint will return elements from parent trees.
 			// We only want to touch each tree once, so we're filtering out nodes on other trees.
-			.filter(nodes => dom.getRootNode(nodes) === root)
+			.filter(nodes => getRootNode(nodes) === root)
 			.reduce((stack, elm) => {
+				// TODO: es-module-utils.isShadowRoot
 				if (axe.utils.isShadowRoot(elm)) {
-					const shadowStack = dom.shadowElementsFromPoint(
+					const shadowStack = shadowElementsFromPoint(
 						nodeX,
 						nodeY,
 						elm.shadowRoot,
@@ -30,7 +33,7 @@ dom.shadowElementsFromPoint = function(nodeX, nodeY, root = document, i = 0) {
 					stack = stack.concat(shadowStack);
 					// filter host nodes which get included regardless of overlap
 					// TODO: refactor multiline overlap checking inside shadow dom
-					if (stack.length && axe.commons.dom.visuallyContains(stack[0], elm)) {
+					if (stack.length && visuallyContains(stack[0], elm)) {
 						stack.push(elm);
 					}
 				} else {
@@ -39,4 +42,6 @@ dom.shadowElementsFromPoint = function(nodeX, nodeY, root = document, i = 0) {
 				return stack;
 			}, [])
 	);
-};
+}
+
+export default shadowElementsFromPoint;
