+++ lib/commons/dom/url-props-from-attribute.js
@@ -1,5 +1,3 @@
-/* global dom */
-
 /**
  * Parse resource object for a given node from a specified attribute
  * @method urlPropsFromAttribute
@@ -7,7 +5,7 @@
  * @param {String} attribute attribute of the node from which resource should be parsed
  * @returns {Object}
  */
-dom.urlPropsFromAttribute = function urlPropsFromAttribute(node, attribute) {
+function urlPropsFromAttribute(node, attribute) {
 	if (!node.hasAttribute(attribute)) {
 		return undefined;
 	}
@@ -52,7 +50,7 @@ dom.urlPropsFromAttribute = function urlPropsFromAttribute(node, attribute) {
 		hash: getHashRoute(parser.hash),
 		filename
 	};
-};
+}
 
 /**
  * Resolve given port excluding default port(s)
@@ -150,3 +148,5 @@ function getHashRoute(hash) {
 
 	return hash;
 }
+
+export default urlPropsFromAttribute;
