+++ lib/commons/dom/visually-contains.js
@@ -1,15 +1,15 @@
-/* global dom */
-
 /**
  * Return the ancestor node that is a scroll region.
  * @param {VirtualNode}
  * @return {VirtualNode|null}
  */
 function getScrollAncestor(node) {
+	// TODO: es-module-utils.getNodeFromTree
 	const vNode = axe.utils.getNodeFromTree(node);
 	let ancestor = vNode.parent;
 
 	while (ancestor) {
+		// TODO: es-module-utils.getScroll
 		if (axe.utils.getScroll(ancestor.actualNode)) {
 			return ancestor.actualNode;
 		}
@@ -86,7 +86,7 @@ function contains(node, parent) {
  * @param {Element} parent
  * @return {boolean} True if node is visually contained within parent
  */
-dom.visuallyContains = function visuallyContains(node, parent) {
+function visuallyContains(node, parent) {
 	const parentScrollAncestor = getScrollAncestor(parent);
 
 	// if the elements share a common scroll parent, we can check if the
@@ -108,4 +108,6 @@ dom.visuallyContains = function visuallyContains(node, parent) {
 	} while (node);
 
 	return false;
-};
+}
+
+export default visuallyContains;
