+++ lib/commons/dom/visually-overlaps.js
@@ -1,5 +1,3 @@
-/* global dom */
-
 /**
  * Checks whether a parent element visually overlaps a rectangle, either directly or via scrolling.
  * @method visuallyOverlaps
@@ -9,7 +7,7 @@
  * @param {Element} parent
  * @return {boolean} True if rect is visually contained within parent
  */
-dom.visuallyOverlaps = function(rect, parent) {
+function visuallyOverlaps(rect, parent) {
 	var parentRect = parent.getBoundingClientRect();
 	var parentTop = parentRect.top;
 	var parentLeft = parentRect.left;
@@ -43,4 +41,6 @@ dom.visuallyOverlaps = function(rect, parent) {
 	}
 
 	return true;
-};
+}
+
+export default visuallyOverlaps;
