+++ lib/commons/table/get-cell-position.js
@@ -1,5 +1,5 @@
-/* global dom */
 import toGrid from './to-grid';
+import findUp from '../dom/find-up';
 
 /**
  * Get the x, y coordinates of a table cell; normalized for rowspan and colspan
@@ -12,8 +12,7 @@ import toGrid from './to-grid';
 function getCellPosition(cell, tableGrid) {
 	var rowIndex, index;
 	if (!tableGrid) {
-		// TODO: es-module-dom.findUp
-		tableGrid = toGrid(dom.findUp(cell, 'table'));
+		tableGrid = toGrid(findUp(cell, 'table'));
 	}
 
 	for (rowIndex = 0; rowIndex < tableGrid.length; rowIndex++) {
