+++ lib/commons/table/get-headers.js
@@ -2,6 +2,8 @@ import isRowHeader from './is-row-header';
 import isColumnHeader from './is-column-header';
 import toGrid from './to-grid';
 import getCellPosition from './get-cell-position';
+import idrefs from '../dom/idrefs';
+import findUp from '../dom/find-up';
 
 /**
  * Loop through the table grid looking for headers and caching the result.
@@ -60,8 +62,7 @@ function traverseForHeaders(headerType, position, tableGrid) {
  */
 function getHeaders(cell, tableGrid) {
 	if (cell.getAttribute('headers')) {
-		// TODO: es-module-dom.idrefs
-		const headers = commons.dom.idrefs(cell, 'headers');
+		const headers = idrefs(cell, 'headers');
 
 		// testing has shown that if the headers attribute is incorrect the browser
 		// will default to the table row/column headers
@@ -70,8 +71,7 @@ function getHeaders(cell, tableGrid) {
 		}
 	}
 	if (!tableGrid) {
-		// TODO: es-module-dom.findUp
-		tableGrid = toGrid(commons.dom.findUp(cell, 'table'));
+		tableGrid = toGrid(findUp(cell, 'table'));
 	}
 	const position = getCellPosition(cell, tableGrid);
