+++ lib/commons/table/get-scope.js
@@ -1,6 +1,6 @@
-/* global dom */
 import toGrid from './to-grid';
 import getCellPosition from './get-cell-position';
+import findUp from '../dom/find-up';
 
 /**
  * Determine if a `HTMLTableCellElement` is a column header, if so get the scope of the header
@@ -30,8 +30,7 @@ function getScope(cell) {
 	} else if (cell.nodeName.toUpperCase() !== 'TH') {
 		return false;
 	}
-	// TODO: es-module-dom.findUp
-	var tableGrid = toGrid(dom.findUp(cell, 'table'));
+	var tableGrid = toGrid(findUp(cell, 'table'));
 	var pos = getCellPosition(cell, tableGrid);
 
 	// The element is in a row with all th elements, that makes it a column header
