+++ lib/commons/table/is-data-cell.js
@@ -1,3 +1,5 @@
+import isValidRole from '../aria/is-valid-role';
+
 /**
  * Determine if a `HTMLTableCellElement` is a data cell
  * @method isDataCell
@@ -12,8 +14,7 @@ function isDataCell(cell) {
 		return false;
 	}
 	const role = cell.getAttribute('role');
-	// TODO: es-module-aria.isValidRole
-	if (axe.commons.aria.isValidRole(role)) {
+	if (isValidRole(role)) {
 		return ['cell', 'gridcell'].includes(role);
 	} else {
 		return cell.nodeName.toUpperCase() === 'TD';
