+++ lib/commons/table/is-data-table.js
@@ -1,4 +1,8 @@
-/* global dom */
+import getRoleType from '../aria/get-role-type';
+import isFocusable from '../dom/is-focusable';
+import findUp from '../dom/find-up';
+import getElementCoordinates from '../dom/get-element-coordinates';
+import getViewportSize from '../dom/get-viewport-size';
 
 /**
  * Determines whether a table is a data table
@@ -13,16 +17,14 @@ function isDataTable(node) {
 	var role = (node.getAttribute('role') || '').toLowerCase();
 
 	// The element is not focusable and has role=presentation
-	// TODO: es-module-dom.isFocusable
-	if ((role === 'presentation' || role === 'none') && !dom.isFocusable(node)) {
+	if ((role === 'presentation' || role === 'none') && !isFocusable(node)) {
 		return false;
 	}
 
 	// Table inside editable area is data table always since the table structure is crucial for table editing
 	if (
 		node.getAttribute('contenteditable') === 'true' ||
-		// TODO: es-module-dom.findUp
-		dom.findUp(node, '[contenteditable="true"]')
+		findUp(node, '[contenteditable="true"]')
 	) {
 		return true;
 	}
@@ -33,8 +35,7 @@ function isDataTable(node) {
 	}
 
 	// Table having ARIA landmark role is data table
-	// TODO: es-module-aria.getRoleType
-	if (commons.aria.getRoleType(role) === 'landmark') {
+	if (getRoleType(role) === 'landmark') {
 		return true;
 	}
 
@@ -171,10 +172,8 @@ function isDataTable(node) {
 
 	// Wide table (more than 95% of the document width) is layout table
 	if (
-		// TODO: es-module-dom.getElementCoordinates
-		dom.getElementCoordinates(node).width >
-		// TODO: es-module-dom.getViewportSize
-		dom.getViewportSize(window).width * 0.95
+		getElementCoordinates(node).width >
+		getViewportSize(window).width * 0.95
 	) {
 		return false;
 	}
