+++ lib/commons/text/accessible-text-virtual.js
@@ -1,4 +1,3 @@
-/* global dom */
 import arialabelledbyText from '../aria/arialabelledby-text';
 import arialabelText from '../aria/arialabel-text';
 import nativeTextAlternative from './native-text-alternative';
@@ -6,6 +5,7 @@ import formControlValue from './form-control-value';
 import subtreeText from './subtree-text';
 import titleText from './title-text';
 import sanitize from './sanitize';
+import isVisible from '../dom/is-visible';
 
 /**
  * Finds virtual node and calls accessibleTextVirtual()
@@ -84,8 +84,7 @@ function shouldIgnoreHidden({ actualNode }, context) {
 		return false;
 	}
 
-	// TODO: es-module-dom.isVisible
-	return !dom.isVisible(actualNode, true);
+	return !isVisible(actualNode, true);
 }
 
 /**
@@ -114,8 +113,7 @@ function prepareContext(virtualNode, context) {
 		context.includeHidden === undefined
 	) {
 		context = {
-			// TODO: es-module-dom.isVisible
-			includeHidden: !dom.isVisible(actualNode, true),
+			includeHidden: !isVisible(actualNode, true),
 			...context
 		};
 	}
