+++ lib/commons/text/form-control-value.js
@@ -1,4 +1,3 @@
-/* global dom */
 import getRole from '../aria/get-role';
 import unsupported from './unsupported';
 import visibleVirtual from './visible-virtual';
@@ -10,6 +9,7 @@ import isAriaListbox from '../forms/is-aria-listbox';
 import isAriaCombobox from '../forms/is-aria-combobox';
 import isAriaRange from '../forms/is-aria-range';
 import getOwnedVirtual from '../aria/get-owned-virtual';
+import isHiddenWithCSS from '../dom/is-hidden-with-css';
 
 const controlValueRoles = [
 	'textbox',
@@ -117,8 +117,7 @@ function ariaTextboxValue(virtualNode) {
 	if (!isAriaTextbox(actualNode)) {
 		return '';
 	}
-	// TODO: es-module-dom.isHiddenWithCSS
-	if (!dom.isHiddenWithCSS(actualNode)) {
+	if (!isHiddenWithCSS(actualNode)) {
 		return visibleVirtual(virtualNode, true);
 	} else {
 		return actualNode.textContent;
