+++ lib/commons/text/label-text.js
@@ -1,6 +1,7 @@
-/* global dom */
 import accessibleTextVirtual from './accessible-text-virtual';
 import accessibleText from './accessible-text';
+import findUpVirtual from '../dom/find-up-virtual';
+import findElmsInContext from '../dom/find-elms-in-context';
 
 /**
  * Return accessible text for an implicit and/or explicit HTML label element
@@ -25,8 +26,7 @@ function labelText(virtualNode, context = {}) {
 
 	const labelContext = { inControlContext: true, ...context };
 	const explicitLabels = getExplicitLabels(virtualNode);
-	// TODO: es-module-dom.findUpVirtual
-	const implicitLabel = dom.findUpVirtual(virtualNode, 'label');
+	const implicitLabel = findUpVirtual(virtualNode, 'label');
 
 	let labels;
 	if (implicitLabel) {
@@ -53,8 +53,7 @@ function getExplicitLabels({ actualNode }) {
 	if (!actualNode.id) {
 		return [];
 	}
-	// TODO: es-module-dom.findElmsInContext
-	return dom.findElmsInContext({
+	return findElmsInContext({
 		elm: 'label',
 		attr: 'for',
 		value: actualNode.id,
