+++ lib/commons/text/label-virtual.js
@@ -1,6 +1,7 @@
-/* global dom */
 import ariaLabelVirtual from '../aria/label-virtual';
 import visible from './visible';
+import getRootNode from '../dom/get-root-node';
+import findUpVirtual from '../dom/find-up-virtual';
 
 /**
  * Gets the visible text of a label for a given input
@@ -23,8 +24,7 @@ function labelVirtual(node) {
 	if (node.actualNode.id) {
 		// TODO: es-module-utils.escapeSelector
 		const id = axe.utils.escapeSelector(node.actualNode.getAttribute('id'));
-		// TODO: es-module-dom.getRootNode
-		doc = axe.commons.dom.getRootNode(node.actualNode);
+		doc = getRootNode(node.actualNode);
 		ref = doc.querySelector('label[for="' + id + '"]');
 		candidate = ref && visible(ref, true);
 		if (candidate) {
@@ -32,8 +32,7 @@ function labelVirtual(node) {
 		}
 	}
 
-	// TODO: es-module-dom.findUpVirtual
-	ref = dom.findUpVirtual(node, 'label');
+	ref = findUpVirtual(node, 'label');
 	candidate = ref && visible(ref, true);
 	if (candidate) {
 		return candidate;
