+++ lib/commons/text/visible-text-nodes.js
@@ -1,3 +1,5 @@
+import isVisible from '../dom/is-visible';
+
 /**
  * Returns an array of visible text virtual nodes
  *
@@ -8,8 +10,7 @@
  * @return {VitrualNode[]}
  */
 function visibleTextNodes(vNode) {
-	// TODO: es-module-dom.isVisible
-	const parentVisible = axe.commons.dom.isVisible(vNode.actualNode);
+	const parentVisible = isVisible(vNode.actualNode);
 	let nodes = [];
 	vNode.children.forEach(child => {
 		if (child.actualNode.nodeType === 3) {
