+++ lib/commons/text/visible-virtual.js
@@ -1,5 +1,5 @@
-/* global dom */
 import sanitize from './sanitize';
+import isVisible from '../dom/is-visible';
 
 /**
  * Returns the visible text of the virtual node
@@ -21,8 +21,7 @@ function visibleVirtual(element, screenReader, noRecursing) {
 			if (child.actualNode.nodeType === 3) {
 				// filter on text nodes
 				const nodeValue = child.actualNode.nodeValue;
-				// TODO: es-module-dom.isVisible
-				if (nodeValue && dom.isVisible(element.actualNode, screenReader)) {
+				if (nodeValue && isVisible(element.actualNode, screenReader)) {
 					return nodeValue;
 				}
 			} else if (!noRecursing) {
